+++ Modules/Helpers/Resources/lang/de/helpers.php
@@ -3,4 +3,5 @@
 return [
     'helper' => 'Helfer',
     'helpers' => 'Helfer',
+    'working_since_days' => 'Arbeitet seit (Tage)',
 ];