+++ Modules/Helpers/Resources/lang/en/helpers.php
@@ -3,4 +3,5 @@
 return [
     'helper' => 'Helper',
     'helpers' => 'Helpers',
+    'working_since_days' => 'Working since (days)',
 ];